refactor: ♻️ Refactor auth errors and optimize hook exports
Updated the `AuthError` function for dynamic error messages, ensuring it
returns a default message if a specific error key isn't provided.
Improved hook module exports by eliminating redundant directory
references and added new hook implementations `useLockScroll`,
`useMetaTag`, and `useMultipleRefs`. The new hooks manage scroll
behavior, assign meta tags, and facilitate multiple refs respectively.
These additions enhance reusability and maintainability of hook-related
functionalities across the application.

Also, made `Slottable` component and `isSlottable` function exportable
to enable reuse elsewhere in the application.
